Remove DataServerBlockReadHanlder
